# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Fraud Detection & Alert System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List user profiles with risk scores
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: riskTier
          in: query
          schema:
            type: string
            enum: [low, medium, high, very_high]
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: User list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedUsers"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user profile with full risk profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /users/{id}/risk-score:
    get:
      operationId: getUserRiskScore
      tags: [users]
      summary: Get current risk score breakdown for a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Risk score
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RiskScore"

  # ── Merchants ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    UserProfile:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        phone:
          type: string
          nullable: true
        riskScore:
          type: number
        riskTier:
          type: string
          enum: [low, medium, high, very_high]
        totalTransactions:
          type: integer
        fraudulentTransactions:
          type: integer
        totalSpend:
          type: number
        blockedAt:
          type: string
          format: date-time
          nullable: true
        flaggedAt:
          type: string
          format: date-time
          nullable: true
        cardCount:
          type: integer
        lastActivityAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, name, email, riskScore, riskTier, totalTransactions, fraudulentTransactions, totalSpend, cardCount, createdAt]
    PaginatedUsers:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/UserProfile"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required: [data, total, page, limit, totalPages]
    RiskScore:
      type: object
      properties:
        userId:
          type: string
        overall:
          type: number
        velocityScore:
          type: number
        geoScore:
          type: number
        deviceScore:
          type: number
        behaviorScore:
          type: number
        mlScore:
          type: number
        factors:
          type: array
          items:
            type: string
        computedAt:
          type: string
          format: date-time
      required: [userId, overall, velocityScore, geoScore, deviceScore, behaviorScore, mlScore, factors, computedAt]
...